requirements.txt
# Daily Digest - Python Dependencies
# Production-ready agent application using Google ADK + Gemini

# ============================================================================
# Core Agent Framework
# ============================================================================
google-adk>=0.3.0
google-generativeai>=0.8.0
google-genai>=0.1.0

# ============================================================================
# Observability (Logging, Tracing, Metrics)
# ============================================================================
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-gcp-trace>=1.6.0

# ============================================================================
# Template Engine (HTML Generation)
# ============================================================================
jinja2>=3.1.2

# ============================================================================
# HTTP Requests
# ============================================================================
requests>=2.31.0
aiohttp>=3.9.0

# ============================================================================
# Data & Time Handling
# ============================================================================
python-dateutil>=2.8.2
pytz>=2023.3

# ============================================================================
# Environment Variables
# ============================================================================
python-dotenv>=1.0.0

# ============================================================================
# Development & Testing
# ============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
ruff>=0.1.5
mypy>=1.7.0

# ============================================================================
# Optional: For enhanced functionality
# ============================================================================
# beautifulsoup4>=4.12.0  # For web scraping if needed
# pandas>=2.1.0           # For data analysis if needed
